colector/requirements.txt
pymongo
celery
kafka-python
psycopg2-binary
datetime
xmltodict
psycopg2-binary
redis
pytest
